refactor: error의 컴포넌트들에 대한 리팩토링 진행
