--- fourtoon-cookie/src/components/error/ErrorComponent/ErrorComponent.styled.ts
@@ -1,8 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-    safeArea: {
-      flex: 1,
-      backgroundColor: '#FFFFFF',
-    },
-});
